Render the crosshair by injecting logic into `updateCameraAndRender` like how modern version of the mod does.
This means the rendering is able to manually render the debug crosshair which requires the private method `orientCamera` in `EntityRenderer`.

Previous `renderWorldDirections` was being overridden which prevented other mods from hooking into it. Now we, cancel the method and manually render the debug crosshair in the same path as the custom crosshair.

It's grim, but a step forward to before.